requirements.txt
Django==5.1.3
django-rest-framework==0.1.0
django-extensions==3.2.3
drf-spectacular==0.27.2
coverage==7.6.8